Extracted ip block reader into a separate class
